# frontend/openapi/backend.json
{"openapi":"3.0.0","paths":{"/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}}},"tags":["health"]}},"/api/applications":{"post":{"operationId":"ApplicationsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateApplicationDTO"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationDTO"}}}}},"tags":["applications"]},"get":{"operationId":"ApplicationsController_findAll","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"skip","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MinimalApplicationDTO"}}}}}},"tags":["applications"]}},"/api/applications/{id}":{"get":{"operationId":"ApplicationsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationDTO"}}}}},"tags":["applications"]},"patch":{"operationId":"ApplicationsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateApplicationDTO"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationDTO"}}}}},"tags":["applications"]},"delete":{"operationId":"ApplicationsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationDTO"}}}}},"tags":["applications"]}}},"info":{"title":"Kraken","description":"Kraken","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"KeyValuePair":{"type":"object","properties":{"key":{"type":"string"},"value":{"type":"string"}},"required":["key","value"]},"CreateApplicationDTO":{"type":"object","properties":{"name":{"type":"string"},"dependsOn":{"type":"array","items":{"type":"string"}},"metadata":{"type":"array","items":{"$ref":"#/components/schemas/KeyValuePair"}}},"required":["name","dependsOn","metadata"]},"MetadataDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"value":{"type":"string"},"applicationId":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","name","value","applicationId","createdAt","updatedAt"]},"DependencyDTO":{"type":"object","properties":{"id":{"type":"string"},"applicationId":{"type":"string"},"dependencyId":{"type":"string"}},"required":["id","applicationId","dependencyId"]},"ApplicationDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"metadata":{"type":"array","items":{"$ref":"#/components/schemas/MetadataDTO"}},"dependsOn":{"type":"array","items":{"$ref":"#/components/schemas/DependencyDTO"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","name","metadata","dependsOn","createdAt","updatedAt"]},"MinimalApplicationDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"dependsOn":{"type":"array","items":{"type":"string"}},"metadata":{"type":"array","items":{"$ref":"#/components/schemas/KeyValuePair"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","name","dependsOn","metadata","createdAt","updatedAt"]},"UpdateApplicationDTO":{"type":"object","properties":{"name":{"type":"string"},"dependsOn":{"type":"array","items":{"type":"string"}},"metadata":{"type":"array","items":{"$ref":"#/components/schemas/KeyValuePair"}}},"required":["name","dependsOn","metadata"]}}}}